requirements.txt
flask
path.py
chardet
zipp